split config to different file to overcome nvhpc limit
